chore(release): update version to 1.1.0 in .release-please-manifest.json and package.json
feat(changelog): add new features and bug fixes for version 1.1.0

In .release-please-manifest.json, update the version of the root project to 1.1.0 to reflect the changes made in this release.

In package.json, update the version to 1.1.0 to match the changes made in this release.

In CHANGELOG.md, add a new section for version 1.1.0 and list the new features and bug fixes that were implemented in this release.