chore(release): update projects/workflows/generator to version 1.0.0
The .release-please-manifest.json file was updated to change the version of "projects/workflows/generator" from 0.0.0 to 1.0.0.

In addition, the CHANGELOG.md file in the "projects/workflows/generator" directory was also updated to include a new entry for version 1.0.0. The entry includes a bug fix for the release please process, which was addressed in commit 57371e2.

Lastly, the version.txt file in the "projects/workflows/generator" directory was updated to reflect the new version 1.0.0.