chore(release): update .release-please-manifest.json and discord generator module version
The .release-please-manifest.json file was updated to include the "projects/workflows/discord/generator/module" with version "1.0.0".

The discord generator module version was updated from "0.0.0" to "1.0.0" in the version.txt file.

This change was made to reflect the addition of the discord module feature and the bug fix related to the release please process.