chore(release): bump incident_management version from 0.0.0 to 1.0.0
The incident_management project has been updated to version 1.0.0. This change was made to address a bug in the release process. The bug fix is documented in the CHANGELOG.md file. Additionally, a revert of a previous commit was made to remove a double entry. The version.txt file has been updated accordingly.