chore(release): update .release-please-manifest.json and discord generator module version
chore(release): update .release-please-manifest.json to include "projects/workflows/discord/generator/module" with version 0.1.0

feat(discord generator module): add discord module and release please process

The discord generator module has been updated to include a new discord module and the release please process. This change was made to enhance the functionality of the module and improve the release process.

chore(version): update discord generator module version to 0.1.0

The version of the discord generator module has been updated to 0.1.0. This change was made to reflect the addition of the new discord module and the release please process.