chore(management-changelog.yml): remove unnecessary comments and add conditional execution for release-related steps
The unnecessary comments in the workflow file have been removed to improve readability.

The following steps in the workflow file have been updated to only execute if releases were created:
- `jobs.release-please`
- `jobs.checks out the repository`
- `jobs.Create SSH key`
- `jobs.Ensure git is using the SSH key`
- `jobs.Setup git`
- `jobs.Initialize Bit`
- `jobs.Resolve component packages from bit.cloud registry`
- `jobs.Bit tag for each path released`

This change ensures that these steps are only executed when releases are created, reducing unnecessary execution and improving workflow efficiency.