chore(package.json): add bug reporting information to package.json
The bug reporting information was added to the package.json file to provide users with a way to report bugs and issues related to the project. The "bugs" field now includes an email address and a URL where users can submit bug reports. This change improves the project's transparency and makes it easier for users to report and track issues.