chore(security-ossar.yml): update branch configuration for push and pull_request events
The branch configuration for the push and pull_request events in the security-ossar.yml workflow file has been updated. The branches now include both "main" and "develop" branches. This change allows the workflow to run on both the "main" and "develop" branches, ensuring that security checks are performed on both branches. Additionally, a schedule has been added to run the workflow every Friday at 3:41 AM.