chore(.gitrepo): update commit and parent references in subrepo configuration
The commit reference in the subrepo configuration has been updated to `5da9368e380b6334cf45b8d0885d2e12edb1c070`. This ensures that the subrepo is synced with the latest changes from the remote repository.

Additionally, the parent reference has been updated to `d2871904c83135118d921f185eb284cba134cdaf`. This indicates the commit from which the subrepo was created.

These changes were made to keep the subrepo configuration up to date and maintain the correct relationship with the remote repository.