chore(release): update logging aspect version from 0.1.1 to 0.1.2
fix(logging): fix build steps

The logging aspect version in the `.release-please-manifest.json` file was updated from 0.1.1 to 0.1.2. This change was made to reflect the correct version of the logging aspect.

In the `CHANGELOG.md` file of the logging aspect, a new release 0.1.2 was added. The release includes a bug fix for build steps, which was addressed in issue #449. The commit hash for the bug fix is c2dbf6e.

The `version.txt` file of the logging aspect was also updated from 0.1.1 to 0.1.2 to match the updated version.