feat(cspell.config.yaml): add cspell configuration file
Add a new cspell.config.yaml file to the project. This file includes the necessary configuration for the cspell spell checker. It also includes metadata such as project information, author details, contributing guidelines, code of conduct, and licensing information.

The configuration file specifies the version of cspell being used, the paths to ignore for spell checking, dictionary definitions, dictionaries to use, and a list of words to ignore.