Chore: secure workflows (#461)
<!-- Thank you for concidering to create a pull request for this
project.

We utilise a bot to analyse the code quality of the pull request. Please
make sure that the bot is happy with your changes before submitting the
pull request.
As an additional benefit of this bot, we can automatically generate a
description for your pull request based on the commit messages.

We would still love you to explain what you have done in your own words,
but if you are in a hurry, you can just submit the pull request without
a description and the bot will generate one for you.

-->

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

- **Chores**
- Enhanced SSH key setup and Git configuration across various workflows
for improved security and efficiency.
- Introduced Workflow Telemetry in multiple workflows for better
monitoring and data collection.
- Updated actions and configurations in workflows related to
development, security, and management to streamline operations and
enhance security measures.
- Adjusted NPM registry configurations for secure and efficient package
management.
- Reorganized steps and updated actions versions to improve workflow
efficiency and security.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>