chore(workspace.jsonc): enable workspace variants and add package configuration for teambit.pkg/pkg (#533)
The workspace.jsonc file was modified to enable workspace variants and
add package configuration for teambit.pkg/pkg. This allows for
publishing packages with specific settings, such as setting the package
manager publish arguments and defining the package.json properties.

<!-- Thank you for concidering to create a pull request for this
project.

We utilise a bot to analyse the code quality of the pull request. Please
make sure that the bot is happy with your changes before submitting the
pull request.
As an additional benefit of this bot, we can automatically generate a
description for your pull request based on the commit messages.

We would still love you to explain what you have done in your own words,
but if you are in a hurry, you can just submit the pull request without
a description and the bot will generate one for you.

-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Activated previously commented-out configuration for workspace
variants.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->