feat: central platform (#536)
<!-- Thank you for concidering to create a pull request for this
project.

We utilise a bot to analyse the code quality of the pull request. Please
make sure that the bot is happy with your changes before submitting the
pull request.
As an additional benefit of this bot, we can automatically generate a
description for your pull request based on the commit messages.

We would still love you to explain what you have done in your own words,
but if you are in a hurry, you can just submit the pull request without
a description and the bot will generate one for you.

-->

---------

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>
Co-authored-by: Eventiva Ltd <132651746+EventivaLtd@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>