feat(generator): add support for harmonyEnvId option in Generator function
feat(template-options): remove unused import of ComponentTemplate

The changes in `generator.ts` add support for a new option `harmonyEnvId` in the `Generator` function. This allows users to specify the harmony environment ID for the generated templates.

The changes in `template-options.ts` remove an unused import of `ComponentTemplate` to clean up the code.