feat(orm/mikro): add Mikro aspect to the project
feat(orm/mikro): add Mikro aspect to the project to maintain Mikro across the platform

feat(orm/mikro): add Mikro aspect to the project to provide a UI and API

feat(orm/mikro): add Mikro aspect to the project to enable composition and extendability

feat(orm/mikro): add Mikro aspect to the project to enable composition and extendability of the platform

feat(orm/mikro): add Mikro aspect to the project to enable composition and extendability of the platform. The MikroAspect class is used to create an instance of the Mikro aspect. It takes an object as a parameter with an 'id' property specifying the ID of the aspect.

feat(orm/mikro): add Mikro aspect to the project to enable composition and extendability of the platform. The MikroAspect class is used to create an instance of the Mikro aspect. It takes an object as a parameter with an 'id' property specifying the ID of the aspect. The MikroAspect class is exported as the default export of the module.

feat(orm/mikro): add Mikro aspect to the project to enable composition and extendability of the platform. The MikroAspect class is used to create an instance of the Mikro aspect. It takes an object as a parameter with an 'id' property specifying the ID of the aspect. The MikroAspect class is exported as the default export of the module. The MikroAspect class is also exported as a named export.

feat(orm/mikro): add Mikro aspect to the project to enable composition and extendability of the platform. The MikroAspect class is used to create an instance of the Mikro aspect. It takes an object as a parameter with an 'id' property specifying the ID of the aspect. The MikroAspect class is exported as the default export of the module. The MikroAspect class is also exported as a named export. The MikroAspect class is used to create an instance of the Mikro aspect. It takes an object as a parameter with an 'id' property specifying the ID of the aspect. The MikroAspect class is exported as the default export of the module. The MikroAspect class is also exported as a named export. The MikroAspect class is used to create an instance of the Mikro aspect.

feat(mikro.node.runtime.ts): add MikroNode class to handle MikroORM initialization and provide EntityManager instance
feat(mikro.node.spec.ts): add test to retrieve MikroNode aspect