feat(customer-support): add CustomerSupport aspect to Atlassian integrations
- Add CHANGELOG.md file to track changes in the customer support aspect
- Add customer-support-config.ts file to define the configuration for the customer support aspect
- Add customer-support.aspect.ts file to create the CustomerSupportAspect class
- Add customer-support.docs.mdx file to provide documentation for the CustomerSupport aspect

feat(customer-support): add Atlassian Customer Support node runtime module

This commit adds the `customer-support.node.runtime.ts` file, which represents an Atlassian Customer Support node. It is responsible for providing support for Atlassian customers through Discord. It depends on the `DiscordjsNode` class and requires a config object of type `CustomerSupportConfig`.

The `CustomerSupportNode` class has a private property `discordjs` that holds an instance of the Discord.js client for interacting with the Discord API. It also has a constructor that takes in the `discordjs` instance and the `config` object.

The commit also adds the `customer-support.node.spec.ts` file, which contains a test for retrieving the aspect.

Lastly, the commit adds the `index.ts` file, which exports the `CustomerSupportAspect` and `CustomerSupportNode` types.

This commit is a part of version 0.1.0 of the Atlassian Customer Support module.