Add User Agreement and Code of Conduct
A detailed User Agreement and Code of Conduct have been added to the projects' workflows directory. This file contains our community guidelines and standards, enforcement policies, and developer certificates, establishing the expected conduct and principles for all participants. All members are asked to adhere to this Code of Conduct for a holistic healthy community, where everyone can feel safe and welcome.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>