chore: Add Eventiva Cooperation Commitment document
This commits adds a new document, "Eventiva Cooperation Commitment", which details the conditions under which our cooperation commitment operates. It includes terms for dealing with violations of the Covered License, establishing our intent for the commitment's irrevocable, binding nature and defining key terms. A reference to the license under which this work is available is also contained.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>