refactor: update Pino logging functions and tests
This commit switches to using the build function in 'pino-pretty' instead of 'PinoPretty'. We have also updated the variable name from 'console' to 'pino' in 'pino.node.spec.ts' and tested its truthiness. This refactoring was done to ensure consistency and correctness in naming and function usage across our logging utilities.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>