feat: add Discord changelog functionality
This commit includes the creation of a new DiscordChangelog class and a sendToDiscord function. The DiscordChangelog class is responsible for handling tasks related to changelogs, while the sendToDiscord function sends these changelogs to a designated Discord channel. It also includes error handling and the ability to customize the fields included in the Discord message.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>