ci: update jsonSchemas.xml and .coderabbit.yaml content
The changes involve updating the path to schema and modifying the instruction format in the .coderabbit.yaml file. The path in jsonSchemas.xml is updated to point to a different location. The format of instructions in .coderabbit.yaml is changed to improve readability by adding line breaks between the instructions.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>