feat: update versions and add new modules in .bitmap and .gitmodules
This commit contains updates to existing versions and the addition of new modules in the .bitmap and .gitmodules files. We have updated the project versions within .bitmap files, and in .gitmodules, we have added a new submodule for the Mattermost Docker. As for the changes in the pnpm-lock.yaml file, we updated versions for numerous dependencies.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>